+++ services/ui/backend/index.js
@@ -33,18 +33,18 @@ const proxies = [
 ];
 
 const proxyRequest = async (req, res, url) => {
-  console.log("PROXY REQUEST ", url);
+  console.log("PROXY REQUEST:", url);
   // const data = await axios({ url: url });
   // console.log(data);
   req
     .pipe(streamableAxios({ url: url }))
     .on("error", (e) => {
-      console.error("error in streamable axios : ", e);
+      console.error("ERROR PROXYING REQUEST:", url, e.code);
       res.sendStatus(500);
     })
     .pipe(res)
     .on("error", (e) => {
-      console.error("error in result : ", e);
+      console.error("ERROR PROXYING REQUEST:", url, e.code);
       res.sendStatus(500);
     });
 };
@@ -63,17 +63,17 @@ proxies.forEach(({ service, path }) => {
 app.get("/check/hotels", (req, res) => {
   const url = `${proxies[1].service}/live`;
   console.log(url);
-  proxyRequest(req, res, url)
+  proxyRequest(req, res, url);
 });
 
 app.get("/check/cars", (req, res) => {
   const url = `${proxies[2].service}/live`;
-  proxyRequest(req, res, url)
+  proxyRequest(req, res, url);
 });
 
 app.get("/check/flights", (req, res) => {
   const url = `${proxies[3].service}/live`;
-  proxyRequest(req, res, url)
+  proxyRequest(req, res, url);
 });
 
 if (process.env.NODE_ENV === "production") {
