+++ services/ui/frontend/src/api/use-info.js
@@ -1,0 +1,15 @@
+import useSWR from "swr";
+import fetcher from "./fetcher";
+
+
+function useInfo(url) {
+  const { loading, data, error } = useSWR(url, fetcher);
+
+  return {
+    loading,
+    error,
+    data,
+  };
+}
+
+export default useInfo;
