+++ services/ui/frontend/src/components/common/DoubleSlider.js
