+++ services/ui/frontend/src/components/common/convert-currency.js
@@ -1,0 +1,5 @@
+function priceConversion(x, { from, to }) {
+  return (x / from) * to;
+}
+
+export default priceConversion;
