--- services/ui/frontend/src/redux/filters/actions.js
@@ -1,9 +1,0 @@
-import { useDispatch } from "react-redux";
-
-export const types = {};
-
-export function useActions() {
-  const dispatch = useDispatch();
-
-  return {};
-}
