--- services/ui/frontend/src/redux/filters/reducer.js
@@ -1,11 +1,0 @@
-import produce from "immer";
-import { types } from "./actions";
-
-const reducer = produce((draft, action) => {
-  switch (action.type) {
-    default:
-      break;
-  }
-});
-
-export default reducer;
